% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discount.R
\name{calculateDiscount}
\alias{calculateDiscount}
\title{Calculate discount from bundles of products}
\usage{
calculateDiscount(items, discounts)
}
\arguments{
\item{items}{data.frame with all items that are purchased. Contains 2 columns:
'Item' with product name and 'Count' with the amount of items purchased}

\item{discounts}{data.frame with all discounts for bundle of two products
purchased together. Contains 3 columns: 'Item1', 'Item2' (product names) and Discount
(applied to every pair of two defined products).}
}
\value{
Function returns a data.frame with 4 columns ("Item", "Count", "Discount"
and "Cost" - same format as in \code{calculateSingleItemCost} function). Each row is
a seperate discount from a bundle of two products. The value in 'Discount' column is
created by combining the names of two products in a bundle. The value in Cost column
is negative to represent that the discount should be deducted from the final cost.
}
\description{
For every set of two products located on the list of discounts function calculates
the minimum amount of items purchased. This value is a total number of bundles of
a specific type and it is multiplied by a discount value for this bundle to get the
total discount value. Those steps are repeated for every defined discount to create
a final table of discounts.
}
