% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discount.R
\name{calculateSingleItemCost}
\alias{calculateSingleItemCost}
\title{Single item cost}
\usage{
calculateSingleItemCost(itemName, itemCount, products)
}
\arguments{
\item{itemName}{name of an item for which the price will be calculated}

\item{itemCount}{total amount of item to purchase}

\item{products}{a data.frame contaning all available products with their regular
prices and special prices}
}
\value{
Function returns a data.frame with 4 columns ("Item", "Count", "Discount"
and "Cost") and 1 or 2 rows (1 row if only regular or special price is used,
2 rows if both of these prices are used)
}
\description{
Function calculates a cost for a specified amount of single item including
regular price and special price for buying a set amount of items
}
