% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getTotalCost}
\alias{getTotalCost}
\title{Total cost calculation}
\usage{
getTotalCost(costTab, decimal = TRUE)
}
\arguments{
\item{costTab}{a data.frame with a list of purchased items. Requires a column called 'Cost'}

\item{decimal}{if TRUE then function will format cost value to have 2 decimal places.}
}
\value{
Function returns total cost of purchased items as numeric (if \code{decimal} is FALSE)
or as character string (when \code{decimal} is TRUE)
}
\description{
Function calculates total cost from a data.frame containing purchased items.
}
