% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{printReceipt}
\alias{printReceipt}
\title{Print receipt}
\usage{
printReceipt(costTab, decimal = TRUE)
}
\arguments{
\item{costTab}{a data.frame with a list of purchased items. Requires a column called 'Cost'}

\item{decimal}{if TRUE then function will format cost value to have 2 decimal places.}
}
\description{
Function prints the final cost table (data.frame) and a cumulative cost including all discounts.
}
