% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{readProducts}
\alias{readProducts}
\title{Read products from file}
\usage{
readProducts(file = "products.csv")
}
\arguments{
\item{file}{path to the file with products list}
}
\value{
Function returns a data.frame with products read from file if the validation
was successfull.
}
\description{
Function runs a validation of a supplied file containing products and throws an error
if it was not successfull.
If the file format was correct then the 'SpecialPrice' column is split into two seperate columns
named 'DiscountItemCount' (amount of items required to apply a special price) and DiscountPrice
(new price which can be applied if enough items are bought).
}
