% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{readPurchaseItems}
\alias{readPurchaseItems}
\title{Read items for purchase from file}
\usage{
readPurchaseItems(products, file = "purchaseItems.csv")
}
\arguments{
\item{file}{path to the file with products list}
}
\value{
Function returns a data.frame with items por purchase read from file if the validation
was successfull.
}
\description{
Function runs a validation of a supplied file containing items to be purchased and throws an error
if it was not successfull.
}
