% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{validateDiscounts}
\alias{validateDiscounts}
\title{Validate input file with discounts}
\usage{
validateDiscounts(products, file = "discounts")
}
\arguments{
\item{file}{path to the file with discount list for buying a pair of products.}
}
\value{
Function returns a list with two elements. Frst element is a logical variable with
the result of validation. If the validation fails the second element on the list is the message
describing what was wrong. If the validation was successfull then the second element will be
a data.frame with discounts read from file.
}
\description{
Function verifies the content of a file which should contain a list of discounts
applicable when buying two different products.
It verifies if the file contains 3 reqauired columns ('Item1', 'Item2' and 'Discount')
and if the 'Discount' column has numeric values.
}
